# Implement parallel rendering