--- a/raingun-lib/Cargo.toml
+++ b/raingun-lib/Cargo.toml
@@ -6,5 +6,6 @@ authors = ["Magnus Bergmark <magnus.bergmark@gmail.com>"]
 [dependencies]
 cgmath = { git = "https://github.com/Mange/cgmath", branch = "serde-09", default-features = false, features = ["eders"] }
 image = "0.12"
+rayon = "0.6"
 serde = "0.9"
 serde_derive = "0.9"