This fixes #8.

With this algorithm in place, I got a much faster render time on my workstation (4 threads).

Here's the results, when I had both implementations present in the code:

## Original implementation

```bash
./target/release/raingun -w 4096 -h 2660 -o 4k-test.jpg examples/test1.yml
```

\# | Render time | Write time
---|-------------|-----------
 1 |       6.19s |      318ms
 2 |       6.20s |      318ms
 3 |       6.23s |      326ms
 4 |       6.16s |      321ms

## Parallel implementation

```bash
PAR=1 ./target/release/raingun -w 4096 -h 2660 -o 4k-test.jpg examples/test1.yml
```

\# | Render time | Write time
---|-------------|-----------
 1 |       1.23s |      319ms
 2 |       1.23s |      319ms
 3 |       1.23s |      319ms
 4 |       1.25s |      318ms

## Conclusions

* The sequential rendering is almost 5 times slower.
* Output image is byte-identical in both cases.